primitives: massive rewrite of headers - change types and hash
This commit changes all of the types in headers from Blake2b which we
were previously using to our own inhouse type of Hash. In the future I
imagine we have custom types of Blake2b that wrap Hash, but that's
something we can implement later as it's not super important right now.

Further, this commit finishes the write functions and the hash functions
for block headers. So now hashing in Rust is fully working, and is
proven with a test of the genesis block that is hard coded into hsd.